pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hormuz-integrated-oil-research"
version = "0.1.0"
description = "Local research system for integrated oil company Hormuz exposure analysis"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Local Research System" }
]
dependencies = [
  "pandas>=2.2.3",
  "numpy>=1.26.4",
  "requests>=2.32.3",
  "httpx>=0.27.2",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.3.0",
  "yfinance>=0.2.61",
  "openpyxl>=3.1.5",
  "XlsxWriter>=3.2.0",
  "matplotlib>=3.9.2",
  "python-docx>=1.1.2",
  "tenacity>=9.0.0",
  "pydantic>=2.10.6",
  "PyYAML>=6.0.2",
  "python-dateutil>=2.9.0.post0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
packages = ["src", "src.utils"]
